Fix histocurve rendering
The x axis used a range of numbers and dates: numbers of the points and
dates for the axis range. This prevented the histocurve from being
rendered.

This changes the graph to use only numbers, but adds a formatter so that
they're still shown as dates. This allows the graph to be rendered
correctly again.